feat: skinned mesh component

+++ NipsEngine/Source/Engine/Render/Resource/VertexTypes.h
@@ -32,7 +32,7 @@ struct FSkeletalMeshVertex
     FVector4 Tangent;
 
     // Bone influence (핵심)
-    uint8 BoneIndices[4];
+    int32 BoneIndices[4]; // -1로 influence 없음 표시, 256개 이상의 Bone을 가진 Mesh 대응
     float BoneWeights[4];
 };
